% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_html.R
\name{basic_html}
\alias{basic_html}
\title{compile with minimal html template}
\usage{
basic_html(...)
}
\arguments{
\item{css}{string, paths to css files}
}
\value{
compiled html document
}
\description{
Useful for creating external htmls in jspsych
}
\examples{


}
