% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions.R
\name{debug_function}
\alias{debug_function}
\title{debug_function}
\usage{
debug_function(name_function)
}
\arguments{
\item{name_function}{name of the function to debug}
}
\value{

}
\description{
Loads the parameters used in the functions present in _targets.R to make debugging easier
}
\examples{
""
}
