% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stimulus_df_to_json.R
\name{stimulus_df_to_json}
\alias{stimulus_df_to_json}
\title{Convert stimulus dataframe to json object}
\usage{
stimulus_df_to_json(df, stimulus, data, pretty = TRUE)
}
\arguments{
\item{df}{dataframe containins stimulus column with html for each stimulus, and additional columns for data codes}

\item{stimulus}{character name of the stimulus column}

\item{data}{character vector names of columns to include as data codes}

\item{pretty}{logical true = pretty format, false=not pretty}
}
\value{
string a json object definition that can supply jsPsych with stimulus information
}
\description{
Convert stimulus dataframe to json object
}
\examples{

# create a stimulus dataframe
 library(dplyr)
 stroop_stim <- data.frame(stimulus = length(16),
    word = rep(c("red","green","blue","yellow"), each=4),
    color = rep(c("red","green","blue","yellow"), 4),
    response = rep(c("r","g","b","y"), 4),
    id = "stroop_stim",
    fontsize = "16pt") \%>\%
  mutate(stimulus = html_stimulus(df = .,
                     html_content = "word",
                     html_element = "p",
                     column_names = c("color","fontsize"),
                     css = c("color", "font-size"),
                     id = "id"))
  head(stroop_stim)

  # convert to json
  stimulus_df_to_json(stroop_stim,"stimulus",c("word","color","response"))

}
