% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions.R
\name{sync_server_local}
\alias{sync_server_local}
\title{sync_server_local}
\usage{
sync_server_local(server_folder, local_folder, direction, only_test = TRUE)
}
\arguments{
\item{server_folder}{folder in server}

\item{local_folder}{folder in local computer}

\item{direction}{local_to_server or server_to_local}

\item{only_test}{TRUE/FALSE, dry run or sync}
}
\value{

}
\description{
Sync files between server and local or viceversa.
}
\examples{
""
}
