% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions.R
\name{update_data}
\alias{update_data}
\title{update_data}
\usage{
update_data(id_protocol, sensitive_tasks = c(""))
}
\arguments{
\item{id_protocol}{id}

\item{sensitive_tasks}{sensitive tasks (files to .vault)}
}
\value{

}
\description{
Update data/id_protocol folder using rsync
}
\examples{
""
}
