% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_to_script.R
\name{write_to_script}
\alias{write_to_script}
\title{Write json to js script}
\usage{
write_to_script(json_object, var_name = NULL)
}
\arguments{
\item{json_object}{string a json object}

\item{var_name}{character default=NULL, set name to declare json object as a variable}
}
\value{
string printed to the html as a <script></script>
}
\description{
To be used in a knitr code chunk with results="asis"
}
\examples{


}
