% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_locally.R
\name{run_locally}
\alias{run_locally}
\title{Run a jspsych experiment}
\usage{
run_locally(
  path = ".",
  show_in = "viewer",
  jspsychr_host = "127.0.0.1",
  jspsychr_port = 8000
)
}
\arguments{
\item{path}{path}

\item{show_in}{string, "viewer" to show in viewer, or "browser" to show in browser}

\item{jspsychr_host}{host defaults to 127.0.0.1}

\item{jspsychr_port}{port defaults to 8000}
}
\description{
Run a jspsych experiment
}
